This is a very basic use of Regex, and Matplotlib
This is a very basic program that takes a list of mock inputs given by students of their favorite primary color. I included various characters in this list of inputs to practice using regular expressions. I then took the data after it was validated using regex, passed the values into counter variables, then put the variables into a dictionary.

I understand that I had imported Pandas, but never used it. I intend on adding to this code and using pandas down the line.